:truck: Refined component/layout project structure
