+++ src/.vuepress/theme/layouts/Portfolio.vue
@@ -1,12 +1,12 @@
 <template>
-	<div class="flex flex-col h-screen">
+	<div class="flex flex-col h-full">
 		<Navbar />
 
-		<main class="mt-16 flex flex-row flex-wrap items-center justify-center">
-			<BlogCard
-				v-for="(post, i) in posts"
+		<main class="mt-16 flex flex-row flex-wrap flex-grow items-center justify-center">
+			<CardPortfolio
+				v-for="(repo, i) in repos"
 				:key="i"
-				:post="post"
+				:repo="repo"
 				data-aos="fade-up"
 				data-aos-duration="1200"
 				:data-aos-delay="i * 50"
@@ -20,7 +20,7 @@
 <script>
 import AOS from 'aos';
 
-import BlogCard from '@theme/components/BlogCard.vue';
+import CardPortfolio from '@theme/components/CardPortfolio.vue';
 import Navbar from '@theme/components/Navbar.vue';
 
 // TODO: Move to theme.styl
@@ -28,24 +28,25 @@ import 'aos/dist/aos.css';
 
 export default {
 	components: {
-		BlogCard,
+		CardPortfolio,
 		Navbar,
 	},
 	data() {
 		return {
-			posts: [],
+			repos: [],
 		};
 	},
-	mounted() {
+	async mounted() {
 		AOS.init({
 			duration: 1000,
 			easing: 'ease-in-out-sin',
 			once: true,
 		});
 
-		this.posts = this.$site.pages
-			.filter((x) => x.path.startsWith('/blog/') && !x.frontmatter.index)
-			.sort((a, b) => new Date(b.frontmatter.date) - new Date(a.frontmatter.date));
+		let res = await fetch('https://api.github.com/users/nurodev/repos');
+		let json = await res.json();
+		let sorted = json.filter(repo => !"NuroDev/NuroDev".includes(repo.full_name)); // Remove the profile README repo
+		this.repos = sorted;
 	},
 };
 </script>
