+++ src/portfolio.md
@@ -1,3 +1,3 @@
 ---
-layout: PortfolioLayout
+layout: Portfolio
 ---
